+++ src/main/java/org/finance/services/MoedaService.java
@@ -112,7 +112,7 @@ public class MoedaService {
         var quantoQueroTotal = calculosCarteira.calcularQuantoQuero(carteiraIdealPorcento, totalCarteira);
         var quantoFaltaTotal = calculosCarteira.calcularQuantoFalta(quantoQueroTotal, valorTotalAtivo);
         var quantidadeQueTenho = aporteService.calcularQuantidadeCompras(moeda.getAportes());
-        var quantidadeQueFaltaTotal = (int) Math.round(calculosCarteira.calcularQuantidadeQueFalta(quantoFaltaTotal, coin.getPrecoDinamico()));
+        var quantidadeQueFaltaTotal = (double) Math.round(calculosCarteira.calcularQuantidadeQueFalta(quantoFaltaTotal, coin.getPrecoDinamico()));
         var comprarOuAguardar = calculosCarteira.informarComprarOuAguardar(carteiraIdealPorcento, carteiraTenhoPorcento);
         var lucroOuPerda = calculosCarteira.calcularLucroOuPerda(valorTotalAtivo, valorTotalAtivoAtual);
